Split the new tree code into its own set of modules
